fingerprint: fix bug
The computation of the fingerprint takes into account the component.
Now the component is first defaulted to the name
and after that, the computation of the fingerpring is performed.

The problem before the current fix was that the component
default was applied before the computation of the fingerprint.

The node version assertion in the tests is now performed correctly.